Feat/fe/inventoryandapi

See merge request s10-final/S10P31B204!107

+++ frontend/lib/models/records/statistic_list.dart
@@ -27,8 +27,8 @@ class StatisticList {
 class Statistic {
   double? time;
   double? dist;
-  int? cal;
-  int? pace;
+  double? cal;
+  double? pace;
   String? date;
   int? count;
